moving keyring logic to trustroot module
