+++ test/unit/internal/test_ctfe.py
@@ -16,7 +16,7 @@ import pretend
 import pytest
 
 from sigstore._internal.ctfe import CTKeyring
-from sigstore._internal.keyring import Keyring, KeyringLookupError
+from sigstore._internal.trustroot import Keyring, KeyringLookupError
 
 
 class TestCTKeyring:
